FQA/requirements.txt
numpy
pytorch
jieba
sklearn
flask
